Disable sticky header to ensure heading anchors link to the right place
As long as we have a sticky header, we can't safely link to an anchor in
the middle of the page (such as in Examples) because the resulting
scroll position will have the target heading be obscured by the sticky
header element.

In github.com we solve this using JavaScript, but the solution is not
easily portable and doesn't always work really well with back-forward
navigation. There are content or CSS workarounds offered on the
internet, but they usually require content changes or sacrificing
accessibility.

The easiest way to allow linking to an anchor on the page is to avoid
having a sticky header.