internal: extract hostnames into const
In an effort to begin supporting multiple GitHub (enterprise) backends,
it is important to extract all hardcoded references to github.com into
a common location. This change uses internal/host.go to that end.
Unfortunately, we were unable to target non Go files, e.g. JSON,
without making them templates. As such, they are still hardcoded.

Due to the complexities of GitHub (enterprise), several consts are
defined. This allows support for custom domains, subdomain isolation,
and other features. It would be nice to boil this down to a simpler api,
but as a first push this is sufficient.