Use git remotes as base repo in `repo view`
We used to do `determineBaseRepo` here, but I think that advanced base
repository resolution to potentially discover the parent repository as
base is not really needed in this command. If I was in a cloned fork, for
example, and did `repo view`, I would expect that my fork's repository
page is viewed instead of that of the parent.