Enforce deprecation of the `api.Client.GraphQL()` method
We want to deprecate manually writing GraphQL strings in favor of using
the `githubv4` library.

The staticcheck linter (invoked as part of `golangci-lint`) should in
theory fail CI on attempted invocations of a deprecated method, but this
seems to only discover one invocation of this method in our project and
not any others. I suspect that staticcheck has a bug around this.