In pr create, default to pushing to base repo if possible
Lots of contributors have write access to the base repository
and don't rely on personal forks. If the base repository is ignored
when selecting the default repository for pr create, it might
have (IMHO) unexpected results like pushing to a new branch in someone
else's fork that gave you push access.