Add custom commands support
When executing `gh <cmd>`, if `<cmd>` is not a built-in command, this
searches `~/.config/gh/gh-commands/*` and in PATH for executables named
`gh-<cmd>` and runs the first one found.

All arguments are forwarded to the executable and the following
environment variables are set:

- GITHUB_TOKEN for API calls;
- GH_BASEREPO with the `OWNER/REPO` pair determined from git remotes.

Additionally, the executable may provide shell completions by respecting
the mode in which the 1st argument is `__complete` and by writing
completion options as separate lines of output.