Simplify BUILD_DATE and make it more portable
Use a simpler approach and add compatibility with the *BSD date command.

Instead of using ifdef, just try with SOURCE_DATE_EPOCH, and if that
fails, fall back to the current date.

On top of that, try with BSD-style and GNU-style flags, to make it work
with both types of systems.

Solves: https://github.com/cli/cli/issues/2578