pr create: Refactor shelling out to `git push`
This introduces a concept of a Git object that is asked for git
operations (only Push for now) from the context of a command. Instead of
stubbing out the `git push` command, the call to `Push()` itself is
stubbed via mock object generated by mockery.

Additionally, this introduces tests that exercise the logic of shelling
out to `git push`, which is now rewritten to ensure synchronous I/O.