Rearrange Conda installation instructions
Originally,  I was thinking of putting Conda in a separate section after
the Windows section, since Conda probably isn't as well known or used.
But reading through the readme again, it seems like arranging it like
the other instructions makes more sense.

I found myself trying to look for the instructions when I first read it
in the MacOS section, but couldn't find the instructions.