Add note about double quotes on non-Unix-like shells
On non-Unix-like shells like Windows Command Prompt, single quotes are
handled differently.  You need to define aliases using double quotes
instead of single quotes.

I added an inline example to illustrate the quotes.  The example is
formatted as inline code blocks in Markdown.  Unfortunately, because Go
uses backticks for raw string literals, I needed to do some rather ugly
string concatenation in order to get the backticks included in the doc
string.

This also rearranges the notes so that the platform specific notes are
at the end of the documentation.