Fixes #4672:  Add --json flag to gh prs check
This pull request adds support for json output the to `gh pr checks` command.
It also adds support for a json format of `*` which is translated to all known fields.
Support for '*' json format affects all commands with a --json flag.

Add support for the --json and --jq flags to `gh pr checks`
Add support for '*' to be passed as fields for the --json flag to allow all fields to be returned