clone: Load "upstream" repo as https
Instead of loading the upstream repo as "git", let's load it as "https"
and avoid users mistakenly pushing to it without even noticing.

Having it as "https" will at least make the users wonder why they have
to add their username and password, giving them the chance to realise a
possible mistake that's about to happen. :-)

Related: #4762

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>