add label to httpmock.Stub
Debugging which stubs go unmatched during testing is kind of a pain.
This change adds an optional label field to httpmock.Stub that can be
used later during verification. If a stub is unmatched, the label is
included in the output.

This changes http.Registry.Register to return a pointer to the stub that
was created. This allows for method chaining when creating a new
httpmock. That means the label is totally optional and doesn't break any
of the existing code.

e.g.
```
reg.Register(
        httpmock.GraphQL(`mutation LabelRemove\b`),
        httpmock.GraphQLMutation(`
        { "data": { "removeLabelsFromLabelable": { "__typename": "" } } }`,
                func(inputs map[string]interface{}) {}),
).SetLabel("label remove")
```