ssh.go: use setup example that should work with any ssh config
The `gh cs ssh` command suggests an example recipe for setting up
openssh integration. The last step appends an `Include` statement to the
user's `~/.ssh/config`.

Unfortunately, this won't always work as intended. If the existing
configuration ends with a `Host` block, the added `Include` statement
will be conditional on whether that block matches.

By preceding the `Include` statement with `Match all`, we can ensure
that it is always evaluated.