Transform passed project names to projectV2 IDs
With this change, project names are now mapped to a pair of ID lists.
The first one contains IDs of old/v1 projects and the second contains
the IDs of next/v2 projects. This allows us to exclude projectV2 IDs
from IssueCreate and PullRequestCreate mutations and in the future,
would allow us to include only projectV2 IDs in the addProjectV2ItemById
mutation.