Allow to add/remove projectsV2 on edit commands
In constrast to regular field update, we track additions and deletions
separately, since project items can be added to and remove from project,
but not in a single mutation. To add item (i.e. issue or PR) to a
project V2, we call AddProjectV2ItemById with the entity's ID and
project's ID. Removing the item from a project is more tricky. The item
known to the project does not correspond to an ID of an issue or PR.
Instead, we firstly fetch all known item IDs that correspond to the
issue or PR, and then remove those from all projects that are requested
to be removed.