Add ARM64 to Darwin releases
Fixes #2426.

In the past, this was blocked by golang/go#42684, but that issue is now
resolved and the Go toolchain is able to sign binaries by default.

I tested running goreleaser both on macOS directly and on Linux, and in
both cases, the resulting gh binary runs on Apple Silicon without issue.