feat: extend --commit flag to match --branch flag behavior
Before, the --commit flag for the `gh browse` command always generated
the `/tree` repository url for the latest commit.

This commit adjusts the `--commit` flag to match the `--branch` flag
behavior. The `--commit` flag now accepts an arbitrary commit sha and
generates the `/tree` url for the repository at that commit. It also
defaults to the latest commit, if no argument is provided to the flag,
to maintain backwards compatibility.