Speed up gh start up time
The tcell library has an `init()` block where it performs some very expensive
indexing that takes 30-40 ms on my machine. This results in fixed overhead for
every gh invocation, even for commands that don't use tcell at all.

This sets an environment variable that instructs tcell to avoid doing that.