gh status: show status even if a comment 404s
Previously, if `gh status` encountered a mention in a comment whose URL
returned a 404, the entire status would fail. However, a 404'd comment
can be benign, for example being part of a repo that is no longer
visible. In this case, other comments should continue to load.

Fixes cli/cli#7872