Fix triage workflow, restore PR support
This commit reverts the triage workflow to restore the previous PR support using the `pull_request_target`, which is more secure as it doesn't allow fork repositories to change the workflow before it runs.

Additionally, this fixes an issue currently causing the workflow to break as cli/cli uses the `discuss` label but github/cli uses the `triage` label.