Fix `repo create` for repo dirs other than `.git`
In `create.go`, rather than invoke `GitDir`, change the `isLocalRepo`
function so it only calls `ToplevelDir` to figure out whether we’re a
work tree at all.
(This is equivalent to `git rev-parse --show-toplevel`.)

Also introduce an `isToplevel` function, which calls `PathFromRoot`
internally.
(This is equivalent to `git rev-parse --show-prefix`.)

This replaces the current implementation, which is based on `--git-dir`
and wrongly assumes that if and only if the current directory is on the
top level, then the output is a) a relative path, which b) points to
the .git directory.

Fixes #8472 [1].

[1]: https://github.com/cli/cli/issues/8472