Fix the description of `gh release upload`
The formatting verb is used incorrectly and produces a misleading
help text for this command:

```
To define a display label for an asset, append text starting with %#% after the file name.
```

instead of the expected:

```
To define a display label for an asset, append text starting with `#` after the file name.
```