Add querySponsorsViaStringManipulation and test
I've added the ability to query the GraphQL API via string manipulation
as well as refactored the original tests for querying for sponsors to
run the whole suite against both methods. That way, as I build out the feature,
I don't need to duplicate test suites for each implementation.

Note, I'm not sure I like this pattern, abstracting out the function name in
the test declaration, for most use cases because the tests can be hard to follow.
However, it seems fine for this exercise as it is conducive to me focusing
on learning the patterns around each method of querying.