Plumb graphql query into `gh sponsors list <username>` command
Calling `gh sponsors list <username>` now queries the graphql endpoint
and prints the resulting sponsor list without formating

Noteably, there is an issue with the plubming for the
querySponsorsViaReflection function that has to do with typing of the
variables passed into the query. I've left a comment above the function
explaining more.