[WIP] Add formating to sponsors list output
Uses the tableprinter to print out the list of a users sponsors with the
SPONSORS header.

WIP notes: the function is currently working as expected (at least for the
happy path), but I'm getting a invalid memory address error for the tests.
I think it's messing up in the iostreams.Test() on line 115, though I
haven't been able to figure out what I've done wrong, yet