Refactor graphQL interface to allow for shallow mocking
There was an over-testing pattern in the code in that we were testing the
functionality of the querying function in both queries_test and list_test.

This was occuring because the current pattern doesn't allow for easy
mocking of the getter function and thus necessitated mocking the http
response for each function. This refactor introduces a Getter function to
the ListOptions interface and provides the factory functions required for
both normal execution and testing of the getter.

This approach has simplified Test_listRun in list_test.go by allowing for
direct mocking of the getter without requiring mocking of the graphQl
response, resulting in a more orthogonal test suite and non-duplicative
testing.