First pass with repo feature detection
This commit is the first pass of having various commands to look at repo features to determine if v1 projects are available or not.

This doesn't delve into creating new tests as there are questions if this is the right approach or not as well as dealing with shurcool graphql query challenges that require refactoring.