Add testing for printHumanIssuePreview around Projects output
To satisfy the first AC in https://github.com/github/cli/issues/490, we
require testing of the output and the feature detector. Seeing as the api
will work for both v1 & v2 endpoints, the first step of is to ensure that
the output isn't showing Projects when no projectCards (a v1 concept) are
included in the issue it is meant to print.