WIP: Debug missing project output for RichIssuePrinter
Currently, calling `gh issue view <issue> --json='projectItems'` does work,
returning the list of v2 projects an issue is associated with. However,
when `gh issue view <issue>` is called, if there are v2 projects assciated
with the issue, it is not printing them as expected.

I'm puzzled as to why, because the unit tests are working for each step of
the logic, but in this commit I've expanded one of the integration test
fixtures to include v2 projects and things are failing. It looks like the
output is correct but the expect is not, which is puzzling to me, but it
alludes to an issue I can't quite figure out.

Anyway, this is my WIP commit for debugging this