WIP: Add project V2 support to gh issue view
The single change to the `defaultFields` array in this file turns on
projectV2 support for this command. The command will now show v2 projects
when compiled and run in the CLI, thereby satisfying all the AC of this
issue.

However, due to the nature of the tests around this command, all of the old
tests for this command break. That is because these tests rely on mocked
http calls, and there is a hidden call required to get v2 project data for
a given issue. The tests fail because the mock is missing.

I've attempted to just add the mock, but simply adding the mock both was
not straight forward (the data structs are embedded in the fetch
functions) and, despite eventually satisfying the http mock registry,
still resulted in test failures.

Thus, this commit has a bunch of broken tests. The follow up work either
needs to figure out how to get the existing tests working properly or
refactor the code such that we are not reliant on these http mocks for
testing.