WIP: refactor findIssue to remove HTTP mocks in tests
I've created an interface for the findIssue method that is called in
viewRun to demonstrate the start of where a refactor may go to address
the issues enumerated by the previous commit. I'll leave it to whoever
picks this work up to decide if they want to continue in this direction
or not.

This direction is coming from the same motivation that led to the
extraction of the rawIssuePrinter and richIssuePrinter